requirements.txt
# local package
-e .

# external requirements
click==8.1.7
coverage==7.4.2
flake8==7.0.0
mariadb=1.1.10
openfoodfacts==0.1.12
pandas==2.2.1
pymysql==1.1.0
python-dotenv>=0.5.1
Sphinx==7.2.6
sqlalchemy==2.0.27
